Module deflate::chained_hash_table

Structs

ChainedHashTable

Tables
Helper struct to let us allocate both head and prev in the same block.

Constants

HASH_MASK

HASH_SHIFT

WINDOW_MASK

WINDOW_SIZE

Functions

create_tables
Create and box the hash chains.

reset_array

update_hash
Returns a new hash value based on the previous value and the next byte

update_hash_conf