Module deflate::compress

Enums

Flush

Functions

block_type_for_length
Determine if the block is long enough for it to be worth using dynamic huffman codes or just Write the data directly.

compress_data_dynamic_n
Inner compression function used by both the writers and the simple compression functions.

flush_to_bitstream
Write all the lz77 encoded data in the buffer using the specified EncoderState, and finish with the end of block code.

write_stored_block