Module deflate::compression_options

Structs

CompressionOptions
A struct describing the options for a compressor or compression function.

Enums

Compression
An enum describing the level of compression to be used by the encoder

SpecialOptions
Enum allowing some special options (not implemented yet)!

Constants

DEFAULT_LAZY_IF_LESS_THAN

DEFAULT_MAX_HASH_CHECKS

DEFAULT_OPTIONS

HIGH_LAZY_IF_LESS_THAN

HIGH_MAX_HASH_CHECKS

MAX_HASH_CHECKS

_HUFFMAN_ONLY