Struct deflate::deflate_state::DeflateState

pub struct DeflateState<W: Write> {
    pub lz77_state: LZ77State,
    pub input_buffer: InputBuffer,
    pub compression_options: CompressionOptions,
    pub encoder_state: EncoderState<W>,
    pub lz77_writer: DynamicWriter,
    pub bytes_written: u64,
}

Fields
lz77_state: LZ77State input_buffer: InputBuffer compression_options: CompressionOptions encoder_state: EncoderState<W> lz77_writer: DynamicWriter bytes_written: u64
Methods

impl<W: Write> DeflateState<W>

fn _new_with_data(input: &[u8],
compression_options: CompressionOptions,
writer: W)
-> DeflateState<W>

fn new(compression_options: CompressionOptions, writer: W) -> DeflateState<W>

fn reset(&mut self, writer: W) -> Result<W>

Resets the status of the decoder, leaving the compression options intact

If flushing the current writer succeeds, it is replaced with the provided one, buffers and status (except compression options) is reset and the old writer is returned.