Struct deflate::encoder_state::EncoderState

pub struct EncoderState<W: Write> {
    huffman_table: HuffmanTable,
    pub writer: LsbWriter<W>,
}

A struct wrapping a writer that writes data compressed using the provided huffman table

Fields
huffman_table: HuffmanTable writer: LsbWriter<W>
Methods

impl<W: Write> EncoderState<W>

fn new(huffman_table: HuffmanTable, writer: W) -> EncoderState<W>

Creates a new encoder state using the provided huffman table and writer

fn write_literal(&mut self, value: u8) -> Result<()>

Encodes a literal value to the writer

fn write_lzvalue2(&mut self, value: LZType) -> Result<()>

fn write_start_of_block(&mut self, fixed: bool, final_block: bool) -> Result<()>

Write the start of a block, returning Err if the write operation fails.

fn write_end_of_block(&mut self) -> Result<()>

fn flush(&mut self) -> Result<()>

Flush the contained writer and it's bitstream wrapper.

fn update_huffman_table(&mut self,
literals_and_lengths: &[u8],
distances: &[u8])
-> Result<(), HuffmanError>

Update the huffman table by generating new huffman codes from length the length values from the provided tables.

fn reset(&mut self, writer: W) -> Result<W>

Reset the encoder state with a new writer, returning the old one if flushing succeeds.