Function deflate::huffman_lengths::calculate_block_length

fn calculate_block_length<F>(
    frequencies: &[u16], 
    dyn_code_lengths: &[u8], 
    get_num_extra_bits: &F
) -> (u64, u64) where
    F: Fn(usize) -> u64, 

Calculate how many bits data with the given frequencies will use when compressed with dynamic code lengths (first return value) and static code lengths (second return value).

Parameters: Frequencies, length of dynamic codes, and a function to get how many extra bits in addition to the length of the Huffman code the symbol will use.