Module deflate::huffman_table

Structs

ExtraBits
A struct representing the data needed to generate the bit codes for a given value and huffman table.

HuffmanCode

HuffmanTable
A structure containing the tables of huffman codes for lengths, literals and distances

Enums

HuffmanError

Constants

END_OF_BLOCK_POSITION

LENGTH_BITS_START

MAX_CODE_LENGTH

MAX_DISTANCE

MAX_MATCH

MIN_DISTANCE

MIN_MATCH

NUM_DISTANCE_CODES

NUM_LENGTH_CODES

NUM_LITERALS_AND_LENGTHS

Statics

BASE_LENGTH

DISTANCE_BASE

DISTANCE_CODES

DISTANCE_EXTRA_BITS

FIXED_CODE_LENGTHS

FIXED_CODE_LENGTHS_DISTANCE

LENGTH_CODE

LENGTH_EXTRA_BITS_LENGTH

Functions

build_length_count_table
Counts the number of values of each length. Returns a tuple containing the longest length value in the table, it's position, and a vector of lengths. Returns an error if table is empty, or if any of the lengths exceed 15.

create_codes

create_codes_in_place
Generats a vector of huffman codes given a table of bit lengths Returns an error if any of the lengths are > 15

get_distance_code
Get the spot in the huffman table for distances distance corresponds to Returns none if the distance is 0, or above 32768

get_distance_code_and_extra_bits

get_length_code
Get the length code that corresponds to the length value

get_length_code_and_extra_bits
Get the code for the huffman table and the extra bits for the requested length.