Function deflate::huffman_table::build_length_count_table

fn build_length_count_table(
    table: &[u8], 
    len_counts: &mut [u16; 16]
) -> (usize, usize)

Counts the number of values of each length. Returns a tuple containing the longest length value in the table, it's position, and fills in lengths in the len_counts slice. Returns an error if table is empty, or if any of the lengths exceed 15.