Struct deflate::huffman_table::HuffmanTable

pub struct HuffmanTable {
    codes: [HuffmanCode; 288],
    distance_codes: [HuffmanCode; 32],
}

A structure containing the tables of huffman codes for lengths, literals and distances

Fields
codes: [HuffmanCode; 288] distance_codes: [HuffmanCode; 32]
Methods

impl HuffmanTable

fn empty() -> HuffmanTable

fn update_from_length_tables(&mut self,
literals_and_lengths: &[u8],
distances: &[u8])
-> Result<(), HuffmanError>

Update the HuffmanTable from the provided length tables. Returns Err if the tables have lengths > 50, the tables are too short, or are otherwise not formed correctly.

fn get_literal(&self, value: u8) -> HuffmanCode

Get the huffman code from the corresponding literal value

fn get_end_of_block(&self) -> HuffmanCode

Get the huffman code for the end of block value

fn get_length_huffman(&self, length: StoredLength) -> (HuffmanCode, HuffmanCode)

Get the huffman code and extra bits for the specified length

returns None if the length is larger than MIN_MATCH or smaller than MAX_MATCH

fn get_distance_huffman(&self,
distance: u16)
-> Option<(HuffmanCode, HuffmanCode)>

Get the huffman code and extra bits for the specified distance

Returns None if distance is 0 or above 32768