Module deflate::length_encode

Structs

ChainNode
A struct representing a node used in the package-merge algorithm

Leaf
A struct representing leaves (same as chainNode, but without the tail)

Enums

EncodedLength
An enum representing the different types in the run-length encoded data used to encode huffman table lenghts

Constants

COPY_PREVIOUS

MIN_REPEAT

REPEAT_ZERO_3_BITS

REPEAT_ZERO_7_BITS

Functions

advance_lookahead

boundary_package_merge
Implementation of boundary package merge algorithm described by Katajainen/Moffat/Turpin in "A Fast and Space-Economical Algorithm for Length-Limited Coding"

encode_lengths
Run-length encodes the lenghts of the values in lenghts according to the deflate specification. This is used for writing the code lenghts for the huffman tables for the deflate stream. Returns a tuple containing a vec of the encoded lengths, and an array describing the frequencies of the different length codes

huffman_lengths_from_frequency

not_max_repetitions

update_out_and_freq

Type Definitions

NodeIndex

WeightType