Function deflate::length_encode::encode_lengths

pub fn encode_lengths<I>(lengths: I) -> Option<(Vec<EncodedLength>, [u16; 19])> where I: Iterator<Item=u8> + Clone

Run-length encodes the lenghts of the values in lenghts according to the deflate specification. This is used for writing the code lenghts for the huffman tables for the deflate stream. Returns a tuple containing a vec of the encoded lengths, and an array describing the frequencies of the different length codes