Function deflate::length_encode::encode_lengths_m

pub fn encode_lengths_m<'a, I>(
    lengths: I, 
    out: &mut Vec<EncodedLength>, 
    frequencies: &mut [u16; 19]
) where
    I: Iterator<Item = &'a u8> + Clone, 

Run-length encodes the lengths of the values in lengths according to the deflate specification. This is used for writing the code lengths for the Huffman tables for the deflate stream.

Populates the supplied array with the frequency of the different encoded length values The frequency array is taken as a parameter rather than returned to avoid excessive memcpy-ing.