Function deflate::length_encode::in_place::enforce_max_code_lengths

pub fn enforce_max_code_lengths(
    num_codes: &mut [u16; 33], 
    num_used: usize, 
    max_len: usize
)

Checks if any of the lengths exceed max_len, and if that is the case, alters the length table so that no codes exceed max_len. This is ported from miniz (which is released as public domain by Rich Geldreich https://github.com/richgel999/miniz/blob/master/miniz.c)

This will not generate optimal (minimim-redundancy) codes, however in most cases this won’t make a large difference.