Function deflate::length_encode::in_place::in_place_lengths

pub fn in_place_lengths(
    frequencies: &[u16], 
    max_len: usize, 
    leaves: &mut Vec<Node>, 
    lengths: &mut [u8]
)

Generate huffman code lengths, using the algorithm described by Moffat and Katajainen in In-Place Calculation of Minimum-Redundancy Codes http://people.eng.unimelb.edu.au/ammoffat/abstracts/mk95wads.html and it's implementation.

This is significantly faster, and seems to generally create lengths that result in length tables that are better compressible than the algorithm used previously. The downside of this algorithm is that it's not length-limited, so if too long code lengths are generated, it might result in a sub-optimal tables as the length-restricting function isn't optimal.