Module deflate::lz77

This module contains functionality for doing lz77 compression of data.

Structs

ChunkState

A struct to keep track of status between calls of process_chunk_lazy

LZ77State

A struct that contains the hash table, and keeps track of where we are in the input data

Enums

LZ77Status

MatchingType

An enum describing whether we use lazy or greedy matching.

ProcessStatus

Status after calling process_chunk.

Constants

DEFAULT_WINDOW_SIZE

MAX_MATCH

MIN_MATCH

NO_RLE

Functions

add_to_hash_table

Add the specified number of bytes to the hash table from the iterators adding start to the position supplied to the hash table.

buffer_full

create_iterators

Create the iterators used when processing through a chunk of data.

lz77_compress_block

Compress a slice with lz77 compression.

match_too_far

If the match is only 3 bytes long and the distance is more than 8 * 1024, it’s likely to take up more space than it would save.

process_chunk

process_chunk_greedy

process_chunk_lazy