Function deflate::lz77::add_to_hash_table

fn add_to_hash_table(
    bytes_to_add: usize, 
    insert_it: &mut Zip<RangeFrom<usize>, Iter<u8>>, 
    hash_it: &mut Iter<u8>, 
    hash_table: &mut ChainedHashTable
)

Add the specified number of bytes to the hash table from the iterators adding start to the position supplied to the hash table.