Function deflate::lz77::lz77_compress_block

pub fn lz77_compress_block<W: OutputWriter>(data: &[u8],
                                            state: &mut LZ77State,
                                            buffer: &mut InputBuffer,
                                            writer: &mut W,
                                            flush: Flush)
                                            -> (usize, LZ77Status)

Compress a slice with lz77 compression.

This function processes one window at a time, and returns when there is no input left, or it determines it's time to end a block.

Returns the number of bytes of the input that were not processed, and a status describing whether there is no input, it's time to finish, or it's time to end the block.