Struct deflate::lz77::LZ77State

pub struct LZ77State {
    hash_table: ChainedHashTable,
    is_first_window: bool,
    is_last_block: bool,
    overlap: usize,
    max_hash_checks: u16,
    lazy_if_less_than: u16,
    matching_type: MatchingType,
}

A struct that contains the hash table, and keeps track of where we are in the input data

Fields
hash_table: ChainedHashTable

Struct containing hash chains that will be used to find matches.

is_first_window: bool

True if this is the first window that is being processed.

is_last_block: bool

Set to true when the last block has been processed.

overlap: usize

How many bytes the last match in the previous window extended into the current one.

max_hash_checks: u16

The maximum number of hash entries to search.

lazy_if_less_than: u16

Only lazy match if we have a match length less than this.

matching_type: MatchingType

Whether to use greedy or lazy parsing

Methods

impl LZ77State

fn from_starting_values(b0: u8,
b1: u8,
max_hash_checks: u16,
lazy_if_less_than: u16,
matching_type: MatchingType)
-> LZ77State

fn _new_warmup(data: &[u8],
max_hash_checks: u16,
lazy_if_less_than: u16,
matching_type: MatchingType)
-> LZ77State

Creates a new LZ77 state, adding the first to bytes to the hash value to warm it up

fn new(max_hash_checks: u16,
lazy_if_less_than: u16,
matching_type: MatchingType)
-> LZ77State

Creates a new LZ77 state

fn reset(&mut self)

Resets the state excluding max_hash_checks and lazy_if_less_than

fn set_last(&mut self)

fn is_last_block(&self) -> bool

fn is_first_window(&self) -> bool