Function deflate::matching::longest_match

pub fn longest_match(data: &[u8],
                     hash_table: &ChainedHashTable,
                     position: usize,
                     prev_length: usize,
                     max_hash_checks: u16)
                     -> (usize, usize)

Try finding the position and length of the longest match in the input data.

Returns

(length, distance from position) If no match is found that was better than prev_length or at all, or we are at the start, the length value returned will be 2.

Arguments:

data: The data to search in. hash_table: Hash table to use for searching. position: The position in the data to match against. prev_length: The length of the previous longest_match check to compare against. max_hash_checks: The maximum number of matching hash chain positions to check.