Module deflate::output_writer

Structs

DynamicWriter

FixedWriter
OutputWriter that doesn't store frequency information

_DummyWriter

Traits

OutputWriter
A trait used by the lz77 compression function to write output. Used to use the same function for compression with both fixed and dynamic huffman codes (When fixed codes are used, there is no need to store frequency information)

Type Definitions

FrequencyType