Trait deflate::output_writer::OutputWriter

pub trait OutputWriter {
    fn write_literal(&mut self, literal: u8);
    fn write_length_distance(&mut self, length: u16, distance: u16);
    fn buffer_length(&self) -> usize;
    fn clear_buffer(&mut self);
    fn get_buffer(&self) -> &[LZValue];
}

A trait used by the lz77 compression function to write output. Used to use the same function for compression with both fixed and dynamic huffman codes (When fixed codes are used, there is no need to store frequency information)

Required Methods

fn write_literal(&mut self, literal: u8)

fn write_length_distance(&mut self, length: u16, distance: u16)

fn buffer_length(&self) -> usize

fn clear_buffer(&mut self)

fn get_buffer(&self) -> &[LZValue]