Struct deflate::write::DeflateEncoder

pub struct DeflateEncoder<W: Write> { /* fields omitted */ }

A DEFLATE encoder/compressor.

A struct implementing a Write interface that takes unencoded data and compresses it to the provided writer using DEFLATE compression.

Examples

use std::io::Write;

use deflate::Compression;
use deflate::write::DeflateEncoder;

let data = b"This is some test data";
let mut encoder = DeflateEncoder::new(Vec::new(), Compression::Default);
encoder.write_all(data).unwrap();
let compressed_data = encoder.finish().unwrap();

Methods

impl<W: Write> DeflateEncoder<W>

fn new<O: Into<CompressionOptions>>(writer: W, options: O) -> DeflateEncoder<W>

Creates a new encoder using the provided compression options.

fn finish(self) -> Result<W>

Encode all pending data to the contained writer, consume this ZlibEncoder, and return the contained writer if writing succeeds.

fn reset(&mut self, w: W) -> Result<W>

Resets the encoder (except the compression options), replacing the current writer with a new one, returning the old one.