Struct deflate::write::ZlibEncoder

pub struct ZlibEncoder<W: Write> { /* fields omitted */ }

A Zlib encoder/compressor.

A struct implementing a Write interface that takes unencoded data and compresses it to the provided writer using DEFLATE compression with Zlib headers and trailers.

Examples

use std::io::Write;

use deflate::Compression;
use deflate::write::ZlibEncoder;

let data = b"This is some test data";
let mut encoder = ZlibEncoder::new(Vec::new(), Compression::Default);
encoder.write_all(data)?;
let compressed_data = encoder.finish()?;

Methods

impl<W: Write> ZlibEncoder<W>

Important traits for ZlibEncoder<W>
impl<W: Write> Write for ZlibEncoder<W>

pub fn new<O: Into<CompressionOptions>>(writer: W, options: O) -> ZlibEncoder<W>

Create a new ZlibEncoder using the provided compression options.

pub fn finish(self) -> Result<W>

Encode all pending data to the contained writer, consume this ZlibEncoder, and return the contained writer if writing succeeds.

pub fn reset(&mut self, writer: W) -> Result<W>

Resets the encoder (except the compression options), replacing the current writer with a new one, returning the old one.

pub fn checksum(&self) -> u32

Return the adler32 checksum of the currently consumed data.