Module deflate::compression_options

This module contains the various options to tweak how compression is performed.

Note that due to the nature of the DEFLATE format, lower compression levels may for some data compress better than higher compression levels.

For applications where a maximum level of compression (irrespective of compression speed) is required, consider using the Zopfli compressor, which uses a specialised (but slow) algorithm to figure out the maximum of compression for the provided data.

Structs

CompressionOptions
A struct describing the options for a compressor or compression function.

Enums

Compression
An enum describing the level of compression to be used by the encoder

SpecialOptions
Enum allowing some special options (not implemented yet)!

Constants

DEFAULT_LAZY_IF_LESS_THAN

DEFAULT_MAX_HASH_CHECKS

DEFAULT_OPTIONS

HIGH_LAZY_IF_LESS_THAN

HIGH_MAX_HASH_CHECKS

MAX_HASH_CHECKS
The maximum number of hash checks that make sense as this is the length of the hash chain.