Module deflate::huffman_table

Structs

ExtraBits
A struct representing the data needed to generate the bit codes for a given value and Huffman table.

HuffmanCode

HuffmanTable
A structure containing the tables of Huffman codes for lengths, literals and distances

Constants

BASE_LENGTH
Base values to calculate the value of the bits in length codes

DISTANCE_BASE

DISTANCE_CODES

END_OF_BLOCK_POSITION
The position in the literal/length table of the end of block symbol

FIXED_CODE_LENGTHS_DISTANCE
Lengths for the distance codes in the pre-defined/fixed Huffman table (All distance codes are 5 bits long)

LENGTH_BITS_START
What number in the literal/length table the lengths start at

LENGTH_CODE
Table used to get a code from a length value (see get_distance_code_and_extra_bits)

LENGTH_EXTRA_BITS_LENGTH
The number of extra bits for the length codes

MAX_CODE_LENGTH
The maximum length of a Huffman code

MAX_DISTANCE

MAX_MATCH

MIN_DISTANCE

MIN_MATCH
The minimum and maximum lengths for a match according to the DEFLATE specification

NUM_DISTANCE_CODES
The number of distance codes in the distance Huffman table

NUM_LENGTH_CODES
The number of length codes in the Huffman table

NUM_LITERALS_AND_LENGTHS
Combined number of literal and length codes

Statics

FIXED_CODE_LENGTHS
Bit lengths for literal and length codes in the fixed Huffman table The Huffman codes are generated from this and the distance bit length table

Functions

build_length_count_table
Counts the number of values of each length. Returns a tuple containing the longest length value in the table, it's position, and fills in lengths in the len_counts slice. Returns an error if table is empty, or if any of the lengths exceed 15.

create_codes_in_place
Generates a vector of Huffman codes given a table of bit lengths Returns an error if any of the lengths are > 15

get_distance_code
Get the spot in the Huffman table for distances distance corresponds to Returns 255 if the distance is invalid. Avoiding option here for simplicity and performance) as this being called with an invalid value would be a bug.

get_distance_code_and_extra_bits

get_length_code
Get the length code that corresponds to the length value Panics if length is out of range.

get_length_code_and_extra_bits
Get the code for the Huffman table and the extra bits for the requested length.

num_extra_bits_for_distance_code
Get the number of extra bits used for a distance code. (Code numbers above NUM_DISTANCE_CODES will give some garbage value.)

num_extra_bits_for_length_code