Module deflate::length_encode::in_place

Structs

Node

Constants

MAX_NUMBER_OF_CODES

NUM_CODES_LENGTH

Functions

enforce_max_code_lengths
Checks if any of the lengths exceed max_len, and if that is the case, alters the length table so that no codes exceed max_len. This is ported from miniz (which is released as public domain by Rich Geldreich https://github.com/richgel999/miniz/blob/master/miniz.c)

in_place_lengths
Generate huffman code lengths, using the algorithm described by Moffat and Katajainen in In-Place Calculation of Minimum-Redundancy Codes http://people.eng.unimelb.edu.au/ammoffat/abstracts/mk95wads.html and it's implementation.

step_1

step_2

validate_lengths

Type Definitions

WeightType