Module deflate::output_writer

Structs

DynamicWriter

Struct that buffers lz77 data and keeps track of the usage of different codes

Enums

BufferStatus

Constants

MAX_BUFFER_LENGTH

The maximum number of literals/lengths in the buffer, which in practice also means the maximum number of literals/lengths output before a new block is started. This should not be larger than the maximum value FrequencyType can represent to prevent overflowing (which would degrade, or in the worst case break compression).

Type Definitions

FrequencyType

The type used for representing how many times a literal, length or distance code has been output to the current buffer. As we are limiting the blocks to be at most 2^16 bytes long, we can represent frequencies using 16-bit values.