Module deflate::zlib

This module contains functionality for generating a zlib header.

The Zlib header contains some metadata (a window size and a compression level), and optionally a block of data serving as an extra dictionary for the compressor/decompressor. The dictionary is not implemented in this library. The data in the header aside from the dictionary doesn’t actually have any effect on the decompressed data, it only offers some hints for the decompressor on how the data was compressed.

Enums

CompressionLevel

Constants

DEFAULT_CINFO

DEFAULT_CM

DEFAULT_CMF

FCHECK_DIVISOR

_DEFAULT_FLEVEL

Functions

add_fcheck

Generate FCHECK from CMF and FLG (without FCKECH )so that they are correct according to the specification, i.e (CMF*256 + FCHK) % 31 = 0. Returns flg with the FCHKECK bits added (any existing FCHECK bits are ignored).

get_zlib_header

Get the zlib header for the CompressionLevel level using the default window size and no dictionary.

write_zlib_header

Write a zlib header with an empty dictionary to the writer using the specified compression level preset.